requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
cohere==5.12.1
qdrant-client==1.12.1
requests==2.32.3
trafilatura==1.12.2
python-dotenv==1.0.1
openai-agents==0.1.4
pydantic==2.10.4